Finished chart and handle network error

+++ app/src/main/java/br/com/angelorobson/application/util/FragmentBase.kt
@@ -29,7 +29,7 @@ open class FragmentBase : Fragment() {
         val builder = AlertDialog.Builder(requireContext())
 
         builder
-            .setMessage(message)
+            .setMessage(message.toInt())
             .setCancelable(false)
             .setPositiveButton("OK") { _, _ -> }
